feat: Navbar and Responsive Grid System
Added an extensible navbar that is rendered in every page, and also
added a hamburger menu to it that does nothing. This needs to be
worked on and updated in the future.

Set up the products to be displayed in a responsive grid system
so that the components itself don't stick to each other. I'm also
attempting to reduce the complexity of the classes that I'm using
since Skeleton UI already implements a lot of it for me.