docs: README refactor
The previous README was simply boilerplate and did not assist too much
with getting started with development and contribution. This commit is
to help change that.

Furthermore, it is to be noted that until the project reaches a stable
state with Auth and the landing pages, the commits will be on the master
branch. Once a stable state has been reached, the master branch will be
locked and the `dev` branch will be utilized for development.