fix: Broken Resource Link and ProductItem model
Did you know that just by not placing a `/` at the end of a URL,
the request immediately becomes a redirect request? Because the
resource doesn't exist at "http://localhost/products" but rather
is located at "http://localhost/products/".

This is a lesson that I don't think I'll forget, and it's annoying
that it took me a day to figure out.