style: Altered the card to utilize more of skeleton UI's internals
The prior card was static to some degree. This one plays a little
better with the framework that I'm using in the backend. Additionally,
the cost and the quantity of the product have been turned into nice
little keyboard-styled text.

I still haven't managed to figure out how to have the button stay at
the bottom rather than show up right after the content. That is
still on a TODO basis.