feat: Toasts
In order to be responsive and give some feedback to the user, I've
set up the Toast alerting system. However, I do want it to be such that
the responsiveness and feel can be conveyed simply through the UI of the
website itself. But, that'll have to wait and I'll have to look into CSS
a little harder for it to happen.